New folder/requirements.txt
flask>=3.0.0
python-telegram-bot>=20.7
python-dotenv>=1.0.0
sqlalchemy>=2.0.23
flask-sqlalchemy>=3.1.1
werkzeug>=3.0.1
requests>=2.31.0
gunicorn>=21.0.0
boto3>=1.34.0
psycopg[binary]>=3.1.0
flask-wtf>=1.2.1
flask-limiter>=3.5.0
